Moved routes to Routes component
